+++ spec/rails_app/app/components/grid/custom_columns.rb
@@ -46,14 +46,14 @@ class Grid::CustomColumns < Netzke::Grid::Base
   def configure(c)
     c.model = "Book"
 
-    c.columns = [
+    c.attributes = [
       :author__first_name, :author__last_name, :author__name, :title, :digitized, :rating, :exemplars, :updated_at
     ]
     super
   end
 
   # This way we'll always have the extra_column, independent of the columns provided in the configuration
-  def columns
+  def attributes
     super + [:extra_column]
   end